test_ai.py
import httpx
import asyncio
import os
from dotenv import load_dotenv

load_dotenv()
api_key = os.getenv("OPENROUTER_API_KEY")
url = "https://openrouter.ai/api/v1/chat/completions"

async def test():
    async with httpx.AsyncClient() as client:
        response = await client.post(
            url,
            headers={"Authorization": f"Bearer {api_key}"},
            json={
                "model": "openai/gpt-4o-mini",
                "messages": [{"role": "user", "content": "Say hello."}]
            }
        )
        print(f"Status: {response.status_code}")
        print(f"Body: {response.text}")

asyncio.run(test())
