requirements.txt
playwright
asyncio
aiosqlite
rich
pydantic
pydantic-settings
python-dotenv
httpx
beautifulsoup4
aiosmtplib
